requirements.txt
aiogram>=3.12.0,<3.13.0
sqlalchemy>=2.0.35
aiosqlite>=0.20.0
python-dotenv>=1.0.1
pydantic>=2.4.1,<2.9.0
pydantic-settings>=2.4.0
fluent.runtime>=0.4.0
Pillow>=10.4.0
punq>=0.7.0
aiohttp>=3.10.5
watchdog>=4.0.0
ffmpeg-python>=0.2.0
python-magic>=0.4.27
